Don't try to publish deleted assets
The default behaviour for ASP.NET web projects is to include everything
in the `wwwroot` directory as content. Since the contents of that
directory are enumerated right at the start of the publishing process,
and in particular before the `gulp clean` command, publishing in a
working directory where assets have previously been built often results
in MSBuild trying to copy files that no longer exist [1] and therefore
raising error MSB3030.

Excluding the `wwwroot` directory from the project's content should
solve this problem.

[1] https://github.com/aspnet/websdk/issues/162