Add request password reset view and controller actions
At the moment the POST action waits for the password reset link to be
generated and sent before returning. This means that the time taken for
the action to respond gives potential attacks a clue as to whether the
email address they submitted belongs to a user.

This could be addressed by doing the work on a background thread, or
better still, through a background job. However I don't feel the risk is
great enough to warrant the introduction of background jobs at this
stage.