Add action filter for ensuring current user is signed out
This filter will be used when there is a need to sign out any
authenticated user before executing the action. This will typically be
the case for actions that can result in a different user being signed
in.

Redirecting to the sign out action instead of signing out the user as
part of the current request is necessary to avoid things like the
anti-forgery token getting out of sync.