Ensure user is signed out before signing in
Users can currently be signed in via both the sign in and the reset
password forms. To ensure a clean sign in, navigating to those forms
will now automatically sign out the current user, if there is one.

Prior to this change the sign in action already had code to sign out the
current user. However, since the sign out was being performed within the
current request pipeline, the anti-forgery token in the sign in form
would be out of sync the current user and rejected on submission.