Add authentication manager method for validating password reset tokens
Given the very small number of tokens expected to be issued each day,
I've chosen to keep the implementation as simple as possible by deleting
all expired tokens every time a password reset token is checked.

If in the future it becomes necessary to make the checking of password
reset tokens more efficient, this method can be updated to only check
the timestamp of the specified token. A separate mechanism (e.g. a
scheduled job) would needed to delete expired tokens from the database.