Don't show required labels on some forms
All of the fields on the sign in, change password and reset password
forms are required so it isn't that helpful (and indeed quite unusual)
to include required field markers on those forms.

This commit therefore adds a tristate `ShowRequiredLabel` property to
the `RequiredLabelTagHelper` that can be used to override the default
visibility of the required label for the field and sets the
corresponding attribute to `false` where necessary.

Despite the fact that there are currently more required fields with the
required label suppressed than not, I still prefer this 'opt-out'
approach to an 'opt-in' one since I think it offers better protection
against accidental omissions.