Ensure menu button is always in top right corner
Previously, on small viewports, the menu button in the top bar would
wrap onto the second line with the tab buttons. This reduced the space
available for the tab buttons and broke the symmetry of the layout.

The commit therefore updates the top bar layout to keep the menu button
in the top right corner at all times. The easiest way to achieve this
was to switch to using a CSS grid layout instead of Flexbox. This means
that the top bar will look slightly broken in Internet Explorer but it
should look fine in modern browsers.