Improve wrapping of tab labels
The breakpoints for top bar tab layout are optimised for a default
font size of 16px. If the user has set a larger default font size in
their browser there are some viewport widths at which the tab labels
need to wrap onto a second line.

Since I was previously using the CSS `line-height` property to
vertically align the labels within a 48 pixel row, when wrapping did
occur there was a huge gap between the two lines. Using Flexbox to
position the tab contents instead resolves that issue.