Adjust top bar breakpoint to allow for longer tab labels
The French tab labels are slightly longer than their English equivalents
which was causing them to wrap onto a second line at some viewport
widths. Increasing the width at which we start trying to display the
tabs on the same row as the logo fixes that (for the default font size
at least).

The impact of this change other users is minimal since both the old and
the new breakpoint sit comfortably between typical mobile portrait
viewports and typical landscape and tablet viewports.