Localize formatting
Previously the `FormatAsHoursAndMinutes` helper method was implemented
as an `IHtmlHelper` extension method. This made it hard to localize as
there was no easy way to provide it with a string localizer.

Moving the method into a separate service solves that problem by
enabling dependency injection and has the added advantage that the
helper can now be used outside of HTML views if necessary.