Fix sign-in/sign-out redirect loop
Previously if the user returned to the sign in page using the browser
back button, response caching by the browser could result in a redirect
loop between the sign in and sign out actions.

There were a couple of ways in which this could occur:

  1) The browser may have cached a response from the sign in action from
     when the user was already authenticated, resulting in the user
     constantly being redirected back to the sign out action even when
     already signed out (with the sign out action then sending the user
     back to the sign in action every time)
  2) The browser may have cached a response from the sign out action
     from when the user was already signed out, resulting in the user
     constantly being redirected back to the sign in action without the
     authentication cookie being cleared (causing the sign in action to
     send the user back to the sign out action every time)

Unfortunately the cookie authentication sign out handler currently
overrides the `Cache-Control` response header to 'no-cache' without
'no-store'. Therefore, instead of using the `ResponseCache` attribute on
the sign out action, I'm having to manually set the `Cache-Control`
header after calling the sign out method on the authentication service.

[1] https://github.com/aspnet/Security/issues/1474