Make it easier to get the current user for a request
Now that the `ValidatePrincipal` method is retrieving the current user
from the database on each request, there isn't any benefit in lazy
loading it.

I've therefore replaced the asynchronous `GetCurrentUser` method on the
authentication manager a synchronous `GetCurrentUser` extension method
on `HttpContext` class that simply acts a convenience method for getting
the current user from the HTTP context items.

The current user is added to the HTTP context items by the
authentication manager through a companion `SetCurrentUser` extension
method. This happens when signing in and when an authenticated principal
is successfully validated, which should be enough to ensure that the
current user is set whenever there is one.