Migrate to Gulp 4
As part of this migration I updated the gulpfile to use:

  + module exports to define public tasks
  + gulp's new `parallel` and `series` APIs to compose tasks
  + gulp's `watch` API to watch for changes to LESS files

I made the latter change because the `gulp-watch-less` package appears
to have been abandoned and doesn't seem to work with Gulp 4. This change
also appears to have fixed a pre-existing issue where the `watchStyles`
task would terminate whenever an error occurred compiling LESS files.