Remove clock from authentication event data provider
This is the first of series of commits that move the responsibility for
setting the timestamps on records out of the data access layer and into
the application/controller layer.

I'm making this change primarily because I'm about to add 'created by'
and 'last modified by' attributes to recipes. These will be easier to
resolve in the controller layer and I'd rather all the values for audit
columns were resolved in the same layer.

Passing timestamps into data access methods also means that multiple
records created or updated in the same transaction can be given matching
timestamps.