Add no-script alternative to top bar menu
Since the top bar popover menu containing the 'Your account' and 'Sign
out' options can only be opened when JavaScript is enabled, users with
JavaScript disabled require an alternative means to access those
options.

This commit therefore adds a page footer that contains the same links as
the top bar menu and is displayed only when JavaScript is disabled.