Add client-side filtering to recipes index
I chose to implement the filtering entirely client-side because:

  1) Showing and hiding recipes as the filter changes makes it easier
     for users to find the recipes they're looking for faster
  2) It was no more work to implement the filtering client-side than it
     would have been server-side
  3) The DOM already contains a row for every recipe in the database so
     there's no additional cost in terms of bytes transferred

The obvious downsides to this approach are scalability and the
JavaScript must be enabled for the filtering to work. However, I'm
not expecting the number of recipes in the database to increase
significantly any time soon and I'm not aware of any users who are
likely to access the site without JavaScript enabled.