Keep vertical scrollbar visible at all times
Previously the vertical scrollbar would only be shown when the page was
longer than the viewport. This meant that the width and centre alignment
of top bar would sometimes change as the user navigated between shorter
and longer pages or, even more noticeably, as the number of matching
rows changed on the all recipes page.

This commit therefore updates the root `html` element style to keep the
vertical scrollbar visible at all times. Note that this doesn't affect
touch devices (and MacOS) where scrollbars are only visible while
scrolling.