Run Gulp `build` as a separate workflow step
Previously we were using a custom MSBuild target to run the Gulp
`rebuild` task when publishing the ASP.NET web project. This was
convenient when packages were being published locally, but we're now
using GitHub actions to publish packages as part of the CI workflow
instead.

This commit therefore updates the CI workflow to run the Gulp `build`
task as a separate step ahead of publishing the ASP.NET web project,
allowing me to remove the custom MSBuild target from the web project.