Change all read-write view model properties to be init-only
View model properties are not expected to change after initialization.
Changing them all to be init-only allows us to enforce that assumption
and will also eventually [1] be enough to satisfy code analysis rule
CA2227 (collection properties should be read only) for view models that
include collections.

[1] The analyzer for CA2227 has been recently updated to skip over
init-only properties. However the change has yet to be included in the
.NET 5.0 SDK so we will still need to suppress it when we migrate to
.NET analyzers with all rules enabled by default:
https://github.com/dotnet/roslyn-analyzers/pull/4463