Simplify `.gitattributes`
Almost all files in this repository should be treated as text so it's
more convenient to instruct Git to store all text files with normalized
line endings and then list the file extensions for the few file types
that should never be treated as text.

There's no need to specify the line ending type as that's taken care of
by the EditorConfig.