Simplify mechanism for rolling back database after each test
Disposing a connection that has an open transaction causes the open
transaction to be automatically disposed first [1].

This commit therefore replaces the relatively cumbersome callback-based
`WithRollback` method with a new `OpenConnectionWithRollback` method
that simply returns a new connection with a transaction already started.
The caller is then responsible for disposing the connection in the
normal way (i.e. via a using statement or declaration).

[1] https://github.com/mysql-net/MySqlConnector/blob/f58a90c17c768c5f88fa44498ecf41f4389afde3/src/MySqlConnector/MySqlConnection.cs#L1012