Update data reader extensions to extend `MySqlDataReader`
Unlike the abstract `DbDataReader` base class we were previously
extending, `MySqlDataReader` provides built-in overloads of the `Get*`
methods that allow the column to be specified by name instead of ordinal
position.

It does not have built-in methods for getting nullable values, however,
nor does it provide a `GetDateTime` overload that allows you to specify
the `DateTimeKind`.

I've therefore removed the `GetInt32` and `GetInt64` extension methods
and renamed the `GetString` extension method so that it doesn't conflict
with the built-in `GetString` overload that shares the same signature.

Finally, because `MySqlDataReader` is sealed, I had to rewrite the tests
to use real `MySqlDataReader` instances instead of mocks.