Reduce frequency of dependabot checks to monthly
Some packages have multiple new releases a week, usually containing
changes that have little or no relevance to this project. Reducing the
frequency of updates to monthly feels like a better trade off between
keeping dependencies up-to-date and not wasting lots of time on upgrades
that are almost immediately superseded.