Configure SendGrid client to throw exceptions
Previously HTTP error responses from the SendGrid API were being ignored
because neither the client nor its callers were checking the response
code.

Configuring the client to raise an exception in the case of an error
response feels like the simplest and most reliable way to ensure that
errors are correctly handled.