Allow exceptions from `SendPasswordResetLink` to propagate
Previously exceptions raised when sending a password reset link would be
caught and logged without any error being displayed to the user. This
was to avoid inadvertently revealing the existence of a user with a
matching email address.

However potentially leaving users waiting for an email that had never
been sent was quite a high price to pay for such a marginal security
benefit, especially since such an attack would only be possible under
error conditions.

I've therefore decided to now let exceptions from the
`SendPasswordResetLink` method propagate as normal so that users see an
error message when the password reset link cannot not be sent.