Switch to recommended analysis mode
Analysis mode options were quite limited in .NET 5. We could either have
very few (`Default`) or no (`AllDisabledByDefault`) rules treated as
build warnings and explicitly opt-in to more rules, or have all rules
enabled and treated as build warnings by default (`AllEnabledByDefault`)
and opt-out of those that were not applicable to our project.

.NET 6 adds a couple of new analysis modes [1] including a `Recommended`
mode that treats all rules that are enabled by default as build
warnings. Switching to this mode results allows us to maintain a similar
level of strictness as before, but with far fewer opt-outs and
suppressions.

[1] https://github.com/dotnet/roslyn/issues/49250#issuecomment-736879786