Handle concurrency exceptions raised when updating a recipe
Previously attempting to save changes to a recipe with a stale base
revision would result in an unhandled exception, ultimately leading to a
generic HTTP 500 server error response and the user losing their
changes.

This commit aims to improve the user experience slightly by re-rendering
the form with a form-level error instead. Any further attempts to submit
the form will still fail with the same error, but the user is at least
able to start a new edit and copy and paste their changes into a new
tab.

Some alternatives I considered included giving users the option to
resubmit the form to forcibly overwrite the previous changes (which
could be achieved by re-rendering the form with the latest base
revision) or removing the concurrency check completely. However, given
that concurrency errors appear to be very rare at the moment (none have
been reported in production since we enabled Bugsnag error reporting
over two months ago), this felt like the most appropriate compromise
between safety, user experience and simplicity of implementation.