Convert recipe edit model to a positional record type
Since this model no longer contains properties with data annotations, it
is no longer affected by the localization issue [1] that was previously
preventing its conversion to a positional record type.

ASP.NET MVC only supports constructor binding for record types with a
single public constructor [2], however, so I had to convert the
constructor used to populate the model from a recipe to a static method
instead.

[1] https://github.com/dotnet/aspnetcore/issues/39551
[2] https://docs.microsoft.com/aspnet/core/mvc/models/model-binding?view=aspnetcore-6.0#constructor-binding-and-record-types