Make model factory methods thread safe
Up to now these methods have only been used in data provider tests that
are all part of the same test collection and therefore never run in
parallel. However we're about to start using these factory methods in
tests that do run in parallel so we now need them to be thread safe.