Make all change password view model properties not nullable
All properties on this model are now required so they don't need to be
nullable.

That said, even though these properties now default to the empty string,
the ASP.NET MVC model binding engine will continue to set them to a null
reference when the corresponding form fields are blank. I don't expect
that to be much of an issue in practice, however, as we always check
that the model is valid before using it.