Make 'Confirm new password' a required field on change password form
So far we've not needed to mark this field as required because the
`Compare` validation ensures that it can only be valid if it matches
the 'New password' field that is required.

However, we're about to make all properties on the change password view
model not-nullable. This will cause the model binding engine to treat
the `ConfirmNewPassword` property as required, with or without the
`Required` attribute. However, without the attribute, the default error
message would be used instead of the custom localized message we've
explicitly specified for all other required fields.

Similarly we explicitly hid [1] the required label for the other two
fields on the form, so we also need to do the same for this one.

[1] 75e33f0b447fa220c6b0613aefd5957948393f76