Mark time zone as required on preferences view model
This property is bound to a select element with no blank item, so there
should be no way for it not to have a value through normal use of the
app. However I still think it's worth annotating it with the required
attribute, partly so that the required status is explicit but also to
ensure that we're not relying solely on the UI for data validation.