Separate recipe attributes out into a separate model
Having the recipe attributes represented by their own model will
eventually allow us to pass them around as a single object all the way
through from the view to the data provider.

Because of an issue [1] with the localization of data annotations, the
properties on the model can't yet be defined using positional
parameters.

[1] https://github.com/dotnet/aspnetcore/issues/39551