Pass recipe attributes model into data provider methods
These methods previously took an entire recipe model as an argument and
silently ignored the properties on the model that weren't relevant for
that operation (for example `Id` when adding a recipe and
`CreatedByUserId` when updating a recipe). This made the contract for
these methods a little more opaque than necessary.

Passing in a recipe attributes model instead (along with separate `id`
and `currentUserId` arguments as necessary) means that these methods no
longer have any redundant inputs.

One thing to note is that, unlike the `CreatedByUserId` and
`ModifiedByUserId` properties previously used to specify the current
user ID, the new `currentUserId` arguments are not nullable. That's
because the recipe model has to represent recipes that were created or
modified before the corresponding columns were added, but we always want
to record the user ID for new recipes and modifications.