Stop autofocusing inputs
I've noticed on Firefox (and potentially other browsers too) that
autofocusing the first input can cause the associated autofill picker to
open on page load, obscuring the page content below. It can also be
quite disorientating for screen reader users [1] to be taken straight
into the page content, and so I now believe the disadvantages of
autofocusing outweigh the benefits.

[1] https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/autofocus#accessibility_considerations