Configure GraphQL endpoint to use token authentication
We want to use to token authentication for requests to the GraphQL
endpoint and cookie authentication for everything else.

Unfortunately there doesn't appear to a simple way to override the
default authentication scheme for a specific endpoint without adding an
`AuthorizeAttribute` [1] which – as the name suggests – also enables
authorization checks. This is problematic because, while most fields in
the GraphQL API will eventually require authorization, some (such as the
authenticate mutation) need to remain accessible without a valid access
token.

This means that we need to leave the authorization checks to Hot
Chocolate's authorization middleware, and so I'm also calling
`AllowAnonymous` on the endpoint builder to add the
`AllowAnonymousAttribute` [2] that suppresses ASP.NET Core's
authorization checks for the endpoint.

[1] https://docs.microsoft.com/en-us/dotnet/api/microsoft.aspnetcore.authorization.authorizeattribute?view=aspnetcore-6.0
[2] https://docs.microsoft.com/en-us/dotnet/api/microsoft.aspnetcore.authorization.allowanonymousattribute?view=aspnetcore-6.0