Dispose dummy connection in test
This should always been a violation of CA1001 (types that own disposable
fields should be disposable) but, due to a bug [1] with the analyzer not
handling target-typed new expressions, hasn't shown up as a compiler
warning.

That bug is now fixed in .NET 7 [2] so I'm fixing the violation before
upgrading.

I also had to seal the test class to avoid violating CA1816 (call
GC.SuppressFinalize correctly).

[1] https://github.com/dotnet/roslyn-analyzers/issues/5834
[2] https://github.com/dotnet/roslyn-analyzers/pull/5880