Temporarily disable CA1852 for top-level statements
As mentioned in the previous commit, .NET 7 includes a new analyzer [1]
that checks for internal types that can be sealed.

Unfortunately that analyzer is currently generating false positives for
generated `Program` classes [2] and therefore has to be suppressed
manually.

A fix [3] has already been committed to the Roslyn Analyzers repo so we
should be able to remove this suppression once that fix has made its way
into an SDK update.

[1] https://github.com/dotnet/roslyn-analyzers/pull/5594
[2] https://github.com/dotnet/roslyn-analyzers/issues/6141
[3] https://github.com/dotnet/roslyn-analyzers/pull/6278