Migrate gulpfile to TypeScript
The latest versions of some of the gulpfile dependencies can only be
imported into ES modules [1]. My original plan was therefore to convert
the gulpfile to an ES module until I realised we could just as easily
migrate it to TypeScript instead [2] and get the added benefits of type
checking.

The TSConfig changes in this commit are necessary to ensure that
`esModuleInterop` is true (through the inherited base configuration [3])
when compiling the gulpfile, while remaining false when compiling the
app bundles.

[1] https://github.com/sindresorhus/del/releases/tag/v7.0.0 and
    https://github.com/sindresorhus/gulp-rev/releases/tag/v10.0.0
[2] https://gulpjs.com/docs/en/getting-started/javascript-and-gulpfiles#transpilation
[3] https://github.com/tsconfig/bases/blob/d699759e29cfd5f6ab0fab9f3365c7767fca9787/bases/recommended.json#L6