Enable linting of *.mjs files
We're going to start using the *.mjs file extension for JavaScript files
that need to be treated as ES modules.

This config change ensures that *.mjs files are linted [1] when the
ESLint is run against a directory (such as when we run `npx eslint .`
locally and in CI) and treated as ES modules by the linter.

[1] By default the ESLint CLI only lints files with the `.js` file
    extension or that match at least one overrides entry in the
    configuration:
    https://eslint.org/docs/latest/use/command-line-interface#--ext