Add generic database fixture
This fixture can be used at class or collection level to provide access
to a test database.

The type parameter `TCollection` is used to generate the database name.
To ensure that only one test is run against each database instance at a
time, a unique `TCollection` should be used for each collection or test
class that uses the fixture. A simple way to guarantee this is to use
the collection or test class itself as the type parameter.