Mark `For` property on tag helper as required
The `For` attribute is used as the selector for this tag helper so it
must have been specified for the tag helper to run. We can therefore
safely mark the corresponding property as required to avoid having to
explicitly check for null or use the null forgiving operator when
reading its value.