Avoid subquery in SQL to get most recently updated recipes
Since the only caller of this method (`HomeController.Index`) already
retrieves the 10 most recently created recipes beforehand, we can
replace the subquery that was previously used to exclude the 10 most
recently created recipes with an explicit list of excluded recipe IDs.

This will make the `GetRecentlyUpdatedRecipes` method much easier to
migrate to Entity Framework Core.