Add model binder to normalize strings prior to validation
Up to now strings have been normalized (trimmed and converted to null if
empty) in the data access layer via the `AddWithStringValue` extension
method on the command parameters collection.

This custom model binder applies the same normalization to all string
values at the point of binding to the ASP.NET MVC model – removing the
need to explicitly apply the normalization to individual properties and
ensuring that normalization occurs prior to model validation.