Use name claim for identity name
Overriding `ClaimsIdentity.NameClaimType` to use the email claim for the
identity name made sense when the identity didn't include a name claim.
However, now that we are including the name claim, so we should revert
to the default and more obvious behaviour of using the name claim for
the identity name.