Recreate claims principal when user revision is stale
With cookie authentication the claims principal is encoded in the
encrypted in the session cookie. The claims principal for
cookie-authenticated requests therefore, previously, didn't reflect any
changes made to the user after sign in.

Since we were already retrieving the user model from the database to
validate the security stamp on each request, there is relatively little
overhead in also using that freshly retrieved model to replace the
claims principal when the user revision is stale.

This ensures that the claims principal now contains the latest user
profile data and preferences on every request.