Configure rules for qualifying references with `this.`
The default configuration for the IDE0003/IDE0009 code style rules is to
prefer that references to instance members are only qualified with
`this.` when strictly necessary. The convention for this repository,
however, is to qualify all references to instance members.

This contradiction has only just become noticeable because, until
recently, the corresponding analysers neither ran on build [1] nor in
the OmniSharp-based C# extension for VS Code.

However, v2.0.320 of the C# VS Code extension (released just a few days
ago) uses a new language server powered by Roslyn instead of OmniSharp
[2] by default, and this appears to have caused the IDE0003/IDE0009 code
style warnings to start showing up in VS Code.

This commit therefore aligns the corresponding code style options with
our repository conventions.

Interestingly setting these options to `true` has also caused IDE0009
(add this or Me qualification) to be enforced on build too. I've no idea
why that is, especially since setting the options to `false` doesn't
seem to cause the corresponding IDE0003 (Remove this or Me
qualification) rule to be raised on build.

[1] https://github.com/dotnet/roslyn/blob/520f605d83795593e4cc48b533fe43b60a92087b/src/Analyzers/Core/Analyzers/EnforceOnBuildValues.cs#L127-L128
[2] https://marketplace.visualstudio.com/items/ms-dotnettools.csharp/changelog