Fix incorrect method reference
The intent of this line of code is to check that the "authenticate"
property contains a JSON representation of the failure payload object.

However, due to a typo we were actually calling `object.Equals` instead.
This was returning false, but the test still passed as we weren't making
any assertions on that value.

This bug was only discovered because the new C# extension in VS Code now
displays IDE0002 (simplify member access) code style warnings.