Split updating principal out of `ChangePassword` method
So that the session used perform a password change isn't immediately
invalidated on the next request, the session cookie must be recreated
with an updated principal that contains the new security stamp.

This was previously handled internally by the `ChangePassword` method,
which was an impediment to separating the password authentication and
session management responsibilities of `IAuthenticationManager` into
dedicated services.

This commit therefore removes the `SignInUser` call from the
`ChangePassword` method and instead introduces a separate
`RefreshPrincipal` method that the controller action is able to use to
refresh the principal on the current session when the password has been
changed successfully.