Pass URL helper into `SendPasswordResetLink`
Previously the `SendPasswordResetLink` method took an `ActionContext`
which it then had to pass to an `IUrlHelperFactory` to obtain a URL
helper.

This approach was more convoluted and restrictive than necessary as any
caller that has access to an `ActionContext` also has access to an
`IUrlHelper` (buy not necessarily the other way round).

This commit therefore updates the `SendPasswordResetLink` method to
receive an `IUrlHelper` directly instead.