Adjust grouping and ordering of fields in tests
This commit re-organises the fields in unit tests in groups with test
helpers are first, then dependencies and their mocks, then finally the
subject.

The fields for mocks have also been renamed to with 'mock' at the end
instead of the start so that the alphabetical ordering of the fields for
dependencies is the same regardless of the use of mocks.