Add data access method for persisting a rehashed password
This method is intended to be used when the password hash needs updating
but underlying (plaintext) password hasn't changed (which is why the
password creation time and security stamp are not updated).

To eliminate any risk of silently reverting a concurrent password
change, the caller is required to specify the revision of the user used
to verify the password. If the specified base revision does not match
the current revision in the database, then the rehashed password is not
persisted.

Unlike other data access methods that get the timestamp directly from
the clock, the method intentionally takes the timestamp as a parameter.
This is because the caller is going to need to know what timestamp was
used for its own return value.