Combine `PasswordAuthenticationService` tests into one test per scenario
I generally prefer to write one test per requirement on the basis that
it's usually the clearest way to document all of the component's
behaviour.

However, for more complex tests like these, I feel that the gymnastics
required to share setup code between multiple tests for the same
scenario impaired the readability of tests so much as to cancel out the
benefits of explicitly testing each requirement individually.

I've therefore chosen to sacrifice the explicitness of individual tests
(and a small amount of code reuse) in favour of clearer test setups.