Don't fail authentication when password hash needs upgrading
Previously the `PasswordAuthenticationService.Authenticate` and
`PasswordAuthenticationService.ChangePassword` methods treated
`PasswordVerificationResult.SuccessRehashNeeded` as equivalent to
`PasswordAuthenticationService.Failed` which would have incorrectly
prevented users with an older password hash from authenticating or
changing their password.

This change simply ensures that `SuccessRehashNeeded` is treated as
equivalent to `Success`. The logic for rehashing the password when
needed will be added in a separate commit.