Rename test methods in `UserDataProviderTests`
Because these tests make real calls to the database and are therefore
comparatively slow to run, `UserDataProviderTests` is already written
using one test per scenario.

Previously those tests were named to describe the expected behaviour,
but that can get cumbersome for tests that check more than a couple of
things.

I've therefore renamed the tests to describe the scenario and instead
used inline comments to outline the requirements tested by the
assertions.