Remove data provider dependencies from `PasswordAuthenticationService`
I’ve yet to find a way of using Entity Framework features that rely on
change tracking in data providers that doesn’t end up tightly coupling
data provider methods to each other and their call sites.

By removing the data provider layer we can write more idiomatic and
efficient EF queries directly in the service, at the expense of no
longer being able to mock database queries in the service’s unit tests.

For the reasons outlined in the EF documentation on testing [1], I’m
finally ready to accept that this is an acceptable compromise.

[1] https://learn.microsoft.com/en-us/ef/core/testing/choosing-a-testing-strategy