Insert security events before writing log messages
Executing a database query is more likely to result in an exception than
writing to the logger. Therefore, to avoid inconsistencies between the
logs and the database, it's safer to insert the security event before
writing to the log. This will become even more important if we start
inserting security events as part of a database transaction.