Add extension methods for retrieving an entity by ID
`FindAsync<T>` is similar to `DbSet<T>.FindAsync` [1] except that it can
be called at the end of an `IQueryable<T>` chain. The trade off is that
this extension method still has to query the database even if a matching
entity is already tracked by the context.

`GetAsync<T>` is also also similar to `FindAsync<T>` except that it
throws a `NotFoundException` instead of returning null when no matching
entity is found.

[1] https://learn.microsoft.com/en-us/dotnet/api/microsoft.entityframeworkcore.dbset-1.findasync