Disable AD0001 as workaround for analyzer bug
The current version of .NET 8.0 contains a bug [1] in
`RouteHandlerAnalyzer` that causes it to throw
`InvalidOperationException` with message 'Failed to resolve well-known
type 'Microsoft.AspNetCore.Builder.EndpointRouteBuilderExtensions' when
the application references any libraries that declare methods in the
`Microsoft.AspNetCore.Builder.EndpointRouteBuilderExtensions` static
class.

A fix [2]  is expected [3] to be included in an upcoming .NET 8 patch
release.

[1] https://github.com/dotnet/aspnetcore/issues/50836
[2] https://github.com/dotnet/aspnetcore/pull/50969
[3] https://github.com/dotnet/aspnetcore/issues/50836#issuecomment-1813301014