Add script to facilitate generating code coverage reports locally
The CI workflow outputs a text summary of the code coverage for the .NET
tests. However, to address any gaps, it's useful to have a more detailed
report showing coverage at method and line level.

With that in mind, I originally intended up modify the CI workflow to
also generate an HTML coverage report and publish it as build artifact.
However, after trying this out a few times, I realised that running a
build and then downloading and unzipping the code coverage report was
more cumbersome than running the tests and report generator locally.

This commit therefore moves the command for generating the coverage
report into a bash script that can be invoked locally as well as in CI.