Exclude Razor views from code coverage report
We currently have no test coverage of the Razor views at all, and any
coverage we add in the future is likely to be through end-to-end tests
that are not necessarily intended to test every branch of the generated
classes.

I therefore think there is very little value in including these views in
the code coverage report and statistics.