Run end-to-end tests in CI
The step to install install Playwright browsers is relatively time
consuming, currently taking approximately 45s to complete on the
GitHub-hosted Ubuntu runner.

However, checking out the code and restoring .NET and Node.js
dependencies takes relatively little time in comparison (around 10s once
cached), so we're able to avoid significantly impacting the overall
duration of the workflow by running the end-to-end tests in a separate,
parallel job.

Another solution that the Playwright docs recommend for avoiding this
time consuming step is to run the tests in a containerized job [1]. I
think that may well be the best option in the longrun, but I've
discounted it for now as it would require us to prepare a Docker image
that contains all of the dependencies needed to run the app (including
.NET and MySQL server).

The `Email__ApiKey` environment variable on the run step is currently
necessary to satisfy the mandatory options validation on app start up.
In practice though we don't what the end-to-end tests sending real
emails so, for now, I've just set this variable to a fake API key that
causes the `EmailSender` to silently fail. In the future I'd like to
provide the option to run the app with a null email sender instead but
that's also beyond the scope of this branch.

Finally the configuration generated by the Playwright installer limits
the number of workers to 1 in CI (as recommended in the Playwright docs
[2] to prioritise stability and reproducibility over speed). As it
happens this limit currently has no effect in practice as GitHub-hosted
Ubuntu runners only have two logical CPU cores [3] and Playwright only
uses half the number of logical CPU cores by default [4].

[1] https://playwright.dev/docs/ci#docker
[2] https://playwright.dev/docs/ci#workers
[3] https://docs.github.com/en/actions/using-github-hosted-runners/about-github-hosted-runners/about-github-hosted-runners#supported-runners-and-hardware-resources
[4] https://playwright.dev/docs/api/class-testconfig#test-config-workers